handle MXSIO_SIGNAL_HALFCLOSED in its own routine
We can get stuck waiting indefinitely for bytes to come in from the
network, and not see the MXSIO_SIGNAL_HALFCLOSED delivered to netstack
by libc when a user's program calls shutdown.

With this CL, "dart /system/apps/hello_fuchsia.dart" now behaves the
same under netstack2 as it does under netstack1.

This also moves reacting to MX_SOCKET_PEER_CLOSED to a routine that is
always listening. This means a program that exits without explicitly
closing its socket will not pin a tcpip.Endpoint. (Fixes NET-18.)

Change-Id: I26627ffd617d14f4e7127047c6ee38896225b4c2