Create netif object to track eth device state.
The purpose is to get a context cancellation function for
all of the state associated with a network interface.
This way when we receive a signal from the *eth.Client that
the ethernet device no longer exists, we can clean up the
routing tables, turn off the DHCP client, and free arena
buffers. (Freeing buffers will come in a subsequent CL.)

This also lets us track per-device DHCP clients.

Change-Id: I6f845d07e353ffcf00bdb5ae7f0144987a2b2a15