Clean up resources when eth devices are removed.
We can use a FIFO_CLOSED signal on the rx handle to determine when
an ethernet device is closed. Plumb this through to netstack, and
release the associated buffers from the shared arena.

With this I can pull and reattach a USB ethernet dongle a dozen
times and get a DHCP address each time.

NET-58 # done

Change-Id: If84980d3c93e3880d7f448cd42faa524b37f26c6